Add script to create index during build time
